# Faster parsing for multiple rows