--- a/src/MySqlConnector/ColumnReaders/BinarySignedInt64ColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/BinarySignedInt64ColumnReader.cs
@@ -1,0 +1,21 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using System.Runtime.InteropServices;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class BinarySignedInt64ColumnReader : IColumnReader
+{
+	internal static BinarySignedInt64ColumnReader Instance { get; } = new BinarySignedInt64ColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return MemoryMarshal.Read<long>(data);
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return checked((int) MemoryMarshal.Read<long>(data));
+	}
+}