--- a/src/MySqlConnector/ColumnReaders/BinaryUnsignedInt8ColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/BinaryUnsignedInt8ColumnReader.cs
@@ -1,0 +1,20 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class BinaryUnsignedInt8ColumnReader : IColumnReader
+{
+	internal static BinaryUnsignedInt8ColumnReader Instance { get; } = new BinaryUnsignedInt8ColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return (byte) data[0];
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return (int) (byte) data[0];
+	}
+}