--- a/src/MySqlConnector/ColumnReaders/BinaryYearColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/BinaryYearColumnReader.cs
@@ -1,0 +1,21 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using System.Runtime.InteropServices;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class BinaryYearColumnReader : IColumnReader
+{
+	internal static BinaryYearColumnReader Instance { get; } = new BinaryYearColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return (int) MemoryMarshal.Read<short>(data);
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return (int) MemoryMarshal.Read<short>(data);
+	}
+}