--- a/src/MySqlConnector/ColumnReaders/BytesColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/BytesColumnReader.cs
@@ -1,0 +1,20 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class BytesColumnReader : IColumnReader
+{
+	internal static BytesColumnReader Instance { get; } = new BytesColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return data.ToArray();
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		throw new InvalidCastException($"Can't convert {columnDefinition.ColumnType} to Int32");
+	}
+}