--- a/src/MySqlConnector/ColumnReaders/GuidBytesColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/GuidBytesColumnReader.cs
@@ -1,0 +1,24 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class GuidBytesColumnReader : IColumnReader
+{
+	internal static GuidBytesColumnReader Instance { get; } = new GuidBytesColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+#if NETCOREAPP2_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER
+		return new Guid(data);
+#else
+		return new Guid(data.ToArray());
+#endif
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		throw new InvalidCastException($"Can't convert {columnDefinition.ColumnType} to Int32");
+	}
+}