--- a/src/MySqlConnector/ColumnReaders/IColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/IColumnReader.cs
@@ -1,0 +1,9 @@
+namespace MySqlConnector.ColumnReaders;
+using MySqlConnector.Protocol.Payloads;
+
+internal interface IColumnReader
+{
+	object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition);
+
+	int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition);
+}