--- a/src/MySqlConnector/ColumnReaders/StringColumnReader.cs
+++ b/src/MySqlConnector/ColumnReaders/StringColumnReader.cs
@@ -1,0 +1,21 @@
+namespace MySqlConnector.ColumnReaders;
+using System.Buffers.Text;
+using System.Text;
+using MySqlConnector.Protocol.Payloads;
+using MySqlConnector.Protocol.Serialization;
+using MySqlConnector.Utilities;
+
+internal sealed class StringColumnReader : IColumnReader
+{
+	internal static StringColumnReader Instance { get; } = new StringColumnReader();
+
+	public object ReadValue(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		return Encoding.UTF8.GetString(data);
+	}
+
+	public int ReadInt32(ReadOnlySpan<byte> data, ColumnDefinitionPayload columnDefinition)
+	{
+		throw new InvalidCastException($"Can't convert {columnDefinition.ColumnType} to Int32");
+	}
+}