… src/MySqlConnector/ColumnReaders/NullColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextBooleanColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextDateTimeColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextDoubleColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextFloatColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextSignedInt16ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextSignedInt32ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextSignedInt64ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextSignedInt8ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextTimeColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextUnsignedInt16ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextUnsignedInt32ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextUnsignedInt64ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TextUnsignedInt8ColumnReader.cs
… src/MySqlConnector/ColumnReaders/TimeSwapBinary16ColumnReader.cs
… src/MySqlConnector/Core/BinaryRow.cs
… src/MySqlConnector/Core/ResultSet.cs
… src/MySqlConnector/Core/Row.cs
… src/MySqlConnector/Core/TextRow.cs
… tests/Benchmark/Benchmark.csproj