@rusher 2023-06-16
force push correction. 
there is something that tickles me: I left it as it was before, but why the columns of type float and double are not allowed to retrieve a value for reader.GetInt32(<pos>) ?