@rusher 2023-06-16
btw, good job compare to MySql.Data 8.0.33 : 
```
|   Method |        Library |      Mean |    Error |
|--------- |--------------- |----------:|---------:|
| getInt64 |     MySql.Data | 163.46 ms | 0.625 ms |
| getInt64 | MySqlConnector |  17.60 ms | 0.256 ms |

```