@bgrainger 2023-06-16
> why the columns of type FLOAT and DOUBLE are not allowed to retrieve a value for reader.GetInt32() since it's allowed for DECIMAL

MySqlConnector doesn't implement lossy type conversions in `GetInt32`. (As a workaround, one can call `Convert.ToInt32(GetValue())` instead and potentially lose data accuracy.)

The exception is for `Decimal` because `SUM` and `AVG` will return a `DECIMAL` value even for integral data: https://github.com/mysql-net/MySqlConnector/issues/54#issuecomment-273275754. This is to avoid unexpected failure when use `GetInt32` to read the `SUM` of an `INT` column.